requirements.txt
gradio==4.36.0
gradio_client==1.0.1
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1
torchaudio==2.3.1
torchvision
matplotlib==3.8.3
click==8.1.7
pillow==10.2.0
python-utils==3.8.2
azure-storage-blob